requirements_cnn.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
numpy>=1.21.0
